RTL Design & SYNTHESIS of a 32-bit MicroProcessor
We had a team of 4 members to design and simulate a 32 bit
microprocessor on a FPGA kit. My role was to monitor the overall code
flow along the various modules and then dump the codes on the FPGA kit.
Parameters involved were:
Target Device : Spartan 3E(FPGA)
Tools Used : Xilinx 9.1
Modelsim SE 5.7f